Det går i ged banke

+++ include/commander_node.h
@@ -1,9 +1,9 @@
 #pragma once
 
 struct Waypoint_s {
-    double jointPositions[6];
-    double velocity[6];
-    double effort[6];
+    float jointPositions[6];
+    float velocity[6];
+    float effort[6];
 };
 
 enum CONTROL_MODE {
